+++ src/main/kotlin/com/cedricziel/idea/concourse/ConcourseIcons.kt
@@ -1,8 +1,11 @@
 package com.cedricziel.idea.concourse
 
 import com.intellij.icons.AllIcons
+import com.intellij.openapi.util.IconLoader
 
 object ConcourseIcons {
+    val CONCOURSE = IconLoader.getIcon("/META-INF/pluginIcon.svg")
+    val CONCOURSE_TASK = IconLoader.getIcon("/META-INF/pluginIcon.svg")
     val INPUT = AllIcons.Graph.ZoomIn
     val OUTPUT = AllIcons.Graph.ZoomOut
     val RESOURCE = AllIcons.Ide.Notification.Gear