+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 ### Added
+- Create distinct file types to detect pipeline and task files
+- Add new file actions and templates
 
 ### Changed